Move step code into it's own file
