+++ app/javascript/packages/document-capture/components/document-capture.tsx
@@ -4,15 +4,8 @@ import { useI18n } from '@18f/identity-react-i18n';
 import { FormSteps, PromptOnNavigate } from '@18f/identity-form-steps';
 import { VerifyFlowStepIndicator, VerifyFlowPath } from '@18f/identity-verify-flow';
 import { useDidUpdateEffect } from '@18f/identity-react-hooks';
-import type { FormStep } from '@18f/identity-form-steps';
 import { getConfigValue } from '@18f/identity-config';
 import { UploadFormEntriesError } from '../services/upload';
-import DocumentsStep from './documents-step';
-import InPersonPrepareStep from './in-person-prepare-step';
-import InPersonLocationPostOfficeSearchStep from './in-person-location-post-office-search-step';
-import InPersonLocationFullAddressEntryPostOfficeSearchStep from './in-person-location-full-address-entry-post-office-search-step';
-import InPersonSwitchBackStep from './in-person-switch-back-step';
-import ReviewIssuesStep from './review-issues-step';
 import UploadContext from '../context/upload';
 import AnalyticsContext from '../context/analytics';
 import Submission from './submission';
@@ -20,8 +13,8 @@ import SubmissionStatus from './submission-status';
 import { RetrySubmissionError } from './submission-complete';
 import SuspenseErrorBoundary from './suspense-error-boundary';
 import SubmissionInterstitial from './submission-interstitial';
-import withProps from '../higher-order/with-props';
 import { InPersonContext } from '../context';
+import { getSteps } from '../steps/getSteps';
 
 interface DocumentCaptureProps {
   /**
@@ -127,74 +120,5 @@ function DocumentCapture({ onStepChange = () => {} }: DocumentCaptureProps) {
     </>
   );
 }
-
-const getReviewStep = (submissionError) => {
-  const { t } = useI18n();
-
-  return {
-    name: 'review',
-    form:
-      submissionError instanceof UploadFormEntriesError
-        ? withProps({
-            remainingAttempts: submissionError.remainingAttempts,
-            isFailedResult: submissionError.isFailedResult,
-            isFailedDocType: submissionError.isFailedDocType,
-            captureHints: submissionError.hints,
-            pii: submissionError.pii,
-            failedImageFingerprints: submissionError.failed_image_fingerprints,
-          })(ReviewIssuesStep)
-        : ReviewIssuesStep,
-    title: t('errors.doc_auth.rate_limited_heading'),
-  };
-};
-
-const getSteps = (submissionError, inPersonURL, inPersonFullAddressEntryEnabled, flowPath) => {
-  const { t } = useI18n();
-  const reviewStep: FormStep = getReviewStep(submissionError);
-
-  if (!submissionError) {
-    return [
-      {
-        name: 'documents',
-        form: DocumentsStep,
-        title: t('doc_auth.headings.document_capture'),
-      },
-    ];
-  }
-
-  const locationStep: FormStep = inPersonFullAddressEntryEnabled
-    ? {
-        name: 'location',
-        form: InPersonLocationFullAddressEntryPostOfficeSearchStep,
-        title: t('in_person_proofing.headings.po_search.location'),
-      }
-    : {
-        name: 'location',
-        form: InPersonLocationPostOfficeSearchStep,
-        title: t('in_person_proofing.headings.po_search.location'),
-      };
-
-  const inPersonProofingSteps: FormStep[] = !inPersonURL
-    ? []
-    : [
-        {
-          name: 'prepare',
-          form: InPersonPrepareStep,
-          title: t('in_person_proofing.headings.prepare'),
-        },
-        locationStep,
-      ];
-
-  const steps: FormStep[] = [reviewStep].concat(inPersonProofingSteps);
-
-  if (flowPath === 'hybrid') {
-    return steps.concat({
-      name: 'switch_back',
-      form: InPersonSwitchBackStep,
-      title: t('in_person_proofing.headings.switch_back'),
-    });
-  }
-  return steps;
-};
 
 export default DocumentCapture;
